Extract DataProviderRequest into a protocol and create an encodable version for added JSON support
