--- Sources/HTTPLib/DataProvider.swift
@@ -1,117 +1,0 @@
-import Foundation
-import HTTPStatusCodes
-import HTTPMethod
-
-#if canImport(FoundationNetworking)
-import FoundationNetworking
-#endif
-
-/// A closure that passes a response to a sent request.
-public typealias DataProviderClosure = (DataProviderResponse) -> Void
-
-/// A request sent to a `DataProvider`
-public struct DataProviderRequest {
-	public let method: HTTPMethod
-	public let url: String
-	public let headers: [String: String]?
-	public let body: Data?
-
-	public init(
-		method: HTTPMethod,
-		url: String,
-		headers: [String: String]? = nil,
-		body: Data? = nil) {
-		self.method = method
-		self.url = url
-		self.headers = headers
-		self.body = body
-	}
-}
-
-/// Data Provider Response.
-///
-/// - Success: Represents a successful response from the provider.
-/// - ProviderError: Represents an error from the provider.
-/// - ProviderUnreachable: Represents a reachability error with the provider.
-public enum DataProviderResponse {
-	case success(data: Data?)
-	case error(code: HTTPStatusCode?, data: Data?)
-	case unreachable
-}
-
-/// DataProvider protocol which allows data fetching logic
-/// to be decoupled from the API instance. This is useful
-/// for mocking results into the API without the need for
-/// external libraries.
-///
-/// @since 1.0
-public protocol DataProvider {
-	/// Sends the specified request to the provider.
-	///
-	/// - parameter request: The request used to send/retrieve the data
-	///   from the provider.
-	/// - parameter callback: The callback called once the request completes
-	///   executing the request.
-	func send(_ request: DataProviderRequest, callback: @escaping DataProviderClosure) throws
-}
-
-public extension DataProvider {
-	@available(macOS 10.15, iOS 13.0.0, tvOS 13.0.0, watchOS 6.0, *)
-	/// Sends the specified request to the provider.
-	/// - Parameter request: The request used to send/retrieve the data
-	///   from the provider.
-	/// - Returns:The data provider's response.
-	func send(_ request: DataProviderRequest) async throws -> DataProviderResponse {
-		try await withCheckedThrowingContinuation { continuation in
-			do {
-				try send(request) { response in
-					continuation.resume(returning: response)
-				}
-			} catch {
-				continuation.resume(throwing: error)
-			}
-		}
-	}
-}
-
-// MARK: - Equality -
-
-extension DataProviderRequest: Equatable {
-	public static func == (left: DataProviderRequest, right: DataProviderRequest) -> Bool {
-		var equalHeaders = false
-		if let leftHeaders = left.headers, let rightHeaders = right.headers {
-			equalHeaders = (leftHeaders == rightHeaders)
-		} else if left.headers == nil && right.headers == nil {
-			equalHeaders = true
-		}
-
-		var equalBody = false
-		if let leftBody = left.body, let rightBody = right.body {
-			equalBody = (leftBody == rightBody)
-		} else if left.body == nil && right.body == nil {
-			equalBody = true
-		}
-
-		return (
-			left.method == right.method &&
-				left.url == right.url &&
-				equalHeaders &&
-				equalBody
-		)
-	}
-}
-
-extension DataProviderResponse: Equatable {
-	public static func == (left: DataProviderResponse, right: DataProviderResponse) -> Bool {
-		switch (left, right) {
-		case (.success(let leftData), .success(let rightData)):
-			return leftData == rightData
-		case (.error(let leftCode, let leftData), .error(let rightCode, let rightData)):
-			return leftCode == rightCode && leftData == rightData
-		case (.unreachable, .unreachable):
-			return true
-		default:
-			return false
-		}
-	}
-}
